refactor hash() option (#2)
Also bumps a bunch of dependencies, and makes some other minor tweaks.

* xo@^0.15.0

* ava@^0.14.0

* lint test.js

* nyc@^6.4.0

* remove trailing whitespace and fix typo in readme

* replace hash option

Both AVA and nyc use the `hash` option to ensure some extra data is included in
the hash, and to store a mapping from the hash to a file. They also need to
(correctly!) use md5-hex to generate the hash.

This commit replaces the `hash` option by a `hashData` function an an `onHash`
callback.

`hashData` receives the original input and additional data and may
return a string, buffer or array of either. This is then included in the hash.

`onHash` is called with the input, additional data, and the resulting hash.
Calling code can use this to maintain mappings.

This is a breaking change.

* include package-hash for caching-transform

Include the package-hash result for caching-transform itself in the resulting
hash. This will ensure the hash changes if caching-transform is updated, e.g. if
it changes its storage format.

* remove unused spies